+++ lib/square_web_payments.dart
@@ -1,11 +1,12 @@
 library square_web_payments;
 
-export 'src/interop/card_details.dart' show CardDetails;
-export 'src/interop/error.dart' show Error;
-export 'src/interop/gift_card_details.dart' show GiftCardDetails;
-export 'src/interop/token_details.dart' show TokenDetails;
-export 'src/interop/token_error_details.dart' show TokenErrorDetails;
-export 'src/interop/token_result.dart' show TokenResult;
+export 'src/models/card_details.dart' show CardDetails;
+export 'src/models/contact.dart' show Contact;
+export 'src/models/error.dart' show Error;
+export 'src/models/gift_card_details.dart' show GiftCardDetails;
+export 'src/models/token_details.dart' show TokenDetails;
+export 'src/models/token_error_details.dart' show TokenErrorDetails;
+export 'src/models/token_result.dart' show TokenResult;
 export 'src/card_payment.dart' show CardPayment;
 export 'src/gift_card_payment.dart' show GiftCardPayment;
 export 'src/payment_method_view.dart' show PaymentMethodView;