+++ lib/src/models/gift_card_details.dart
@@ -1,0 +1,8 @@
+/// https://developer.squareup.com/reference/sdks/web/payments/objects/GiftCardDetails
+class GiftCardDetails {
+  final String? type;
+
+  const GiftCardDetails({required this.type});
+
+  Map<String, dynamic> toJson() => type == null ? {} : {'type': type};
+}